chore: #49 데이터 분류
