+++ src/pages/recruit/RecruitDetailPage/RecruitDetailPage.tsx
@@ -1,21 +1,17 @@
 import React, { useCallback, useState } from 'react';
 import SRecruit from './RecruitDetailPage.styled';
-import { Tag, Icon, ApplyInfomation, ApplyInput, ApplySubmit } from '../../../components';
+import {
+	Tag,
+	Icon,
+	ApplyInfomation,
+	ApplyInput,
+	ApplySubmit,
+	informationList,
+	role,
+} from '../../../components';
 import ColorMatching from '../../../utils/ColorMatching';
 import { useRecoilValue } from 'recoil';
 import { applyStepState } from '../../../atom';
-
-interface RequiredInformation {
-	title: string;
-	content: string;
-}
-
-interface Role {
-	role: string;
-	max: number;
-	current: string[];
-	specs: string[];
-}
 
 interface Comment {
 	// id: number;
@@ -26,54 +22,6 @@ interface Comment {
 type ComponentProps = {
 	[key: number]: JSX.Element;
 };
-
-const informationList: RequiredInformation[] = [
-	{
-		title: '범위',
-		content: '교내',
-	},
-	{
-		title: '유형',
-		content: '프로젝트',
-	},
-	{
-		title: '진행',
-		content: '오프라인',
-	},
-	{
-		title: '분야',
-		content: '개발',
-	},
-	{
-		title: '기간',
-		content: '23.10.23 - 24.02.14',
-	},
-	{
-		title: '태그',
-		content: '응소실,Spring,C#,Window Form',
-	},
-];
-
-const role: Role[] = [
-	{
-		role: '디자이너',
-		max: 2,
-		current: ['Valverde'],
-		specs: ['Figma'],
-	},
-	{
-		role: '프론트엔드 개발자',
-		max: 2,
-		current: ['Bellinham', 'Son'],
-		specs: ['TypeScript', 'React'],
-	},
-	{
-		role: '백엔드 개발자',
-		max: 2,
-		current: ['Rice', 'Salah'],
-		specs: ['Node.js', 'Spring'],
-	},
-];
 
 const RecruitDetailPage = () => {
 	// const location = useLocation();
